+++ FindPOLine/hooks/useOrderLine/useOrderLine.test.js
@@ -1,0 +1,34 @@
+import { QueryClient, QueryClientProvider } from 'react-query';
+
+import { renderHook, waitFor } from '@folio/jest-config-stripes/testing-library/react';
+import { useOkapiKy } from '@folio/stripes/core';
+
+import { orderLine } from 'fixtures';
+import { useOrderLine } from './useOrderLine';
+
+const queryClient = new QueryClient();
+
+// eslint-disable-next-line react/prop-types
+const wrapper = ({ children }) => (
+  <QueryClientProvider client={queryClient}>
+    {children}
+  </QueryClientProvider>
+);
+
+describe('useOrderLine', () => {
+  it('should fetch order line by id', async () => {
+    useOkapiKy
+      .mockClear()
+      .mockReturnValue({
+        get: () => ({
+          json: () => Promise.resolve(orderLine),
+        }),
+      });
+
+    const { result } = renderHook(() => useOrderLine(orderLine.id), { wrapper });
+
+    await waitFor(() => expect(result.current.isLoading).toBeFalsy());
+
+    expect(result.current.orderLine).toEqual(orderLine);
+  });
+});